+++ yucca/image_processing/transforms/Spatial.py
@@ -5,7 +5,7 @@ From: https://stackoverflow.com/questions/59738230/apply-rotation-defined-by-eul
 
 import numpy as np
 from scipy.ndimage import map_coordinates
-from yucca.image_processing.matrix_ops import (
+from yucca.functional.matrix_ops import (
     create_zero_centered_coordinate_matrix,
     deform_coordinate_matrix,
     Rx,
